#30- 커뮤 홈 작업 및 반응형 적용
laout.scss에 스크롤바 없애는 작업 추가, 소모임 컴포넌트로 따로 뻄,
