+++ src/scss/components/_layout.scss
@@ -1,7 +1,13 @@
 .content-container {
-    position: absolute;
-    top: 0;
-    width: 100%;
-    height: calc(100vh - 100px);
-    overflow-y: auto;
+  position: absolute;
+  top: 0;
+  width: 100%;
+  height: calc(100vh - 100px);
+  overflow-y: auto;
+  scrollbar-width: none;
+  -ms-overflow-style: none;
+}
+
+.content-container::-webkit-scrollbar {
+  display: none;
 }
