+++ src/scss/page/_community.scss
@@ -10,6 +10,10 @@
   align-items: flex-start;
   gap: 7.5rem;
   justify-content: center;
+  @media screen and (max-width: 360px) {
+    gap: 60%;
+    align-items: center;
+  }
 }
 .community-header {
   display: flex;
@@ -118,8 +122,9 @@
   flex-shrink: 0;
 }
 .community-search-command-box {
+  position: relative;
   display: flex;
-  width: 16.875rem;
+  width: 80%;
   height: 2.5rem;
   padding: 0.625rem 0.9375rem;
   align-items: center;
@@ -129,12 +134,50 @@
   border: 0.0625rem solid var(--calendar_-date_text, #cdcdcd);
   margin: 0.3125rem 0rem 0rem 0.375rem;
 }
+.community-search-input {
+  width: 100%;
+  border: none;
+  outline: none;
+  padding: 0 1rem;
+  font-size: 0.8rem;
+  border-radius: 3.125rem;
+}
 .community-search-command-button {
+  position: absolute;
+  right: 0.625rem;
   width: 1.25rem;
   height: 1.25rem;
-  flex-shrink: 0;
-  float: right;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
+.community-home-made-group {
+  display: block;
+}
+.community-home-subheader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 16px;
+  padding: 10px 16px;
+}
+
+.community-home-subheader-title {
+  font-size: 18px;
+  font-weight: bold;
+}
+
+.community-home-create-group {
+  color: #000;
+  font-family: Pretendard;
+  font-size: 13px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  cursor: pointer;
+  font-size: 16px;
+}
+
 .tabs {
   display: flex;
   width: 100%;
